+++ interpreter/Variable.java
@@ -9,7 +9,7 @@ public class Variable implements Expression {
 		this.value = value;
 	}
 
-	public int result(byte[] stats) {
+	public int result(int[] stats) {
 		return value.result(stats);
 	}
